refactored out the user factory
